==> test_notebooks/02_hedonic_index.ipynb <==
# 02 – Hedonic Index Legacy Notebook

For reproducibility this now calls inal_code.hedonic_indices.

from final_code.hedonic_indices import generate_hedonic_tables
tables = generate_hedonic_tables(save=True)
tables['overall'].head()
