==> test_notebooks/02_hedonic_indices.ipynb <==
# 02 – Hedonic Indices (Test Harness)

This notebook proxies the production hedonic export logic.

from final_code.hedonic_indices import generate_hedonic_tables
tables = generate_hedonic_tables(save=True)
list(tables.keys())

tables['mesh_full'].head()
